fix cycles
